pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gaceta-oficial-crawler"
version = "0.1.0"
description = "Extract edition data from Gaceta Oficial HTML index files and output as JSON"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coloredlogs>=15.0.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    "beautifulsoup4>=4.14.2",
    "pydeps>=3.0.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.1",
]
dev = [
    "mypy>=1.5.0",
    "ruff>=0.14.4",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "samples*", "output*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--verbose --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["domain", "application", "adapters"]
omit = [
    "*/__init__.py",
    "*/test_*",
    "*/tests/*",
    "adapters/main_cli.py",
    "infrastructure/*",
    "main.py",
    "scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
show_missing = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[dependency-groups]
dev = [
    "lxml>=6.0.2",
    "types-requests>=2.32.4.20250913",
]

[project.scripts]
check-all = "scripts.check_all:main"
gaceta = "adapters.main_cli:main"
